third_party/blink/web_tests/platform/generic/svg/animations/animate-to-low-prio-frozen-expected.txt

PASS rectWidth is > 100
PASS 200 is >= rectWidth
PASS successfullyParsed is true

TEST COMPLETE
